api/requirements.txt
Flask==3.0.3
mysqlclient==2.2.4
flask_mysqldb==2.0.0
Werkzeug==3.0.2
Flask-Cors==3.0.10
mysql-connector-python==8.4.0
Flask-JWT-Extended==4.6.0